additions to build script: abort if version is not updated

+++ build.sh
@@ -1,14 +1,21 @@
 #!/bin/sh
 
-VERSION="1.0.0-RC1"
+VERSION="1.0.0-RC2"
+
+# Version check
+truever="$(go run . --version)"
+if [[ $truever != "EasyCopy v"$VERSION ]]; then
+	echo "Version mismatch, should be \"EasyCopy v$VERSION\", but is \"$truever\""
+	echo "Aborting"
+	exit
+fi
 
 mkdir out pkg -p
-rm out/* pkg/* -f
+rm out/* -f
 
 echo "linux/amd64"
 rm pkg/* -f
-go build -v .
-cp easy-copy pkg/
+go build -v -o pkg/easy-copy .
 cp shared/* pkg/
 cp doc/* pkg/
 cp LICENSE pkg/
@@ -22,8 +29,7 @@ cp pkg/"easycopy-$VERSION-linux-amd64.tar.zst" out/
 echo ""
 echo "linux/arm"
 rm pkg/* -f
-GOOS=linux GOARCH=arm go build -v .
-cp easy-copy pkg/
+GOOS=linux GOARCH=arm go build -v -o pkg/easy-copy .
 cp shared/* pkg/
 cp doc/* pkg/
 cp LICENSE pkg/
@@ -37,10 +43,11 @@ cp pkg/"easycopy-$VERSION-linux-arm.tar.zst" out/
 echo ""
 echo "windows/amd64"
 rm pkg/* -f
-GOOS=windows GOARCH=amd64 CGO_ENABLED=1 CC=x86_64-w64-mingw32-gcc go build -v .
-cp easy-copy.exe pkg/
+GOOS=windows GOARCH=amd64 CGO_ENABLED=1 CC=x86_64-w64-mingw32-gcc go build -v -o pkg/easy-copy.exe.
 cp LICENSE pkg/
 cd pkg/
 zip "easycopy-$VERSION-windows-amd64.zip" *
 cd ..
 cp pkg/"easycopy-$VERSION-windows-amd64.zip" out/
+
+rm pkg/*
